fix(package): publish the SysTest XML parser the server imports at runtime
The installed server would not start. Every `npm i -g d365fo-mcp` on 1.16.1 died
on load with

  ERR_MODULE_NOT_FOUND: Cannot find module
  '<pkg>/dist/eval/oracle/systest.js' imported from
  '<pkg>/dist/tools/sdlc/sysTestRunner.js'

taking the whole tool surface with it.

run_systest_class reads per-method results out of the XML document
SysTestConsole writes, and since the TDD-loop work (0887d9f) it imported that
parser from the eval oracle. But src/eval/** is a dev-only tree that
package.json deliberately keeps out of the tarball ("!dist/eval/**"), so the
published sysTestRunner.js imports a file that is not there. Confirmed with
`npm pack --dry-run`: the tarball ships sysTestRunner.js and zero dist/eval
entries.

Nothing in the repo could see it. typecheck, lint and all 5,916 tests resolve
against src/, where the excluded tree is present; the breakage exists only in
the tarball, which is why a green suite shipped it.

parseSysTestXml moves to src/tools/sdlc/sysTestXml.ts — beside the runner that
loads it, inside a published tree — and src/eval/oracle/systest.ts re-exports
it, so the dependency runs eval -> tools and never back. The `files` list is
unchanged; the eval tree stays out of the package, as intended.

The gate checks the closure, not this one pair: publishedFiles.test.ts now walks
every specifier in every packed dist/**/*.js and fails on any that resolves
outside the tarball, so the next tree cut out of the package is covered too.
Verified as a negative control — restoring the old import makes it report
exactly `dist/tools/sdlc/sysTestRunner.js -> ../../eval/oracle/systest.js`.
Comments are stripped first: src/bridge/index.ts documents its own import path
in a JSDoc example, and a doc line is not a module edge.
